[selector-reports] Tweak test-id rewrite test output
